Skin-Disease-Detection-master/requirements.txt
torch==2.2.0
torchvision==0.17.2
Pillow>=10.0.0
numpy==1.26.4
flask>=2.0
